refactor: Test 코드 관련 클래스 패키지 정리
