+++ src/test/kotlin/com/flab/ticketing/common/utils/QRCodeGeneratorTest.kt
@@ -1,6 +1,6 @@
 package com.flab.ticketing.common.utils
 
-import com.flab.ticketing.common.UnitTest
+import com.flab.ticketing.testutils.UnitTest
 import com.google.zxing.BinaryBitmap
 import com.google.zxing.MultiFormatReader
 import com.google.zxing.client.j2se.BufferedImageLuminanceSource
