+++ src/test/kotlin/com/flab/ticketing/testutils/PerformanceTestDataGenerator.kt
@@ -1,4 +1,4 @@
-package com.flab.ticketing.common
+package com.flab.ticketing.testutils
 
 import com.flab.ticketing.common.entity.Region
 import com.flab.ticketing.performance.entity.Performance
