+++ src/test/kotlin/com/flab/ticketing/testutils/conditions/NonCiEnvironment.kt
@@ -1,4 +1,4 @@
-package com.flab.ticketing.common.conditions
+package com.flab.ticketing.testutils.conditions
 
 import io.kotest.core.annotation.EnabledCondition
 import io.kotest.core.spec.Spec
@@ -7,6 +7,6 @@ import kotlin.reflect.KClass
 class NonCiEnvironment : EnabledCondition {
 
     override fun enabled(kclass: KClass<out Spec>): Boolean {
-        return System.getenv("CI") != "true"
+        return !(System.getenv("CI") == "true" || System.getenv("GITHUB_ACTIONS") == "true")
     }
 }
